Use math.js to parse formula for the cell
With { type: "module" } on the Worker, we can use ES2015 modules in
Workers (with almost 100% browser compatibility) allowing us to include
a parser instead of having to write one ourselves.

Picked math.js based on update frequency and number of Github stars.

They do provide a page on their security concerns:

    https://mathjs.org/docs/expressions/security.html

but for the purposes of the tech test I think this is fine.