Support formula editing
Now we pass the formula all the way through to the worker, and have that
deal with swapping out the Cell references for their actual values.

I found this way was easier, because we could then bring the formula all
the way back and update both values at once.

This required a Custom Editor so we can edit a value that's different to
what is displayed in the grid when it's readonly. It also meant updating
the datatype for the cells to be an object so we can store the
underlying formula as well as the result computed by the Worker.